Move 'addSocialMedia', 'removeSocialMedia', 'handleSocialMedia' to separate files.
