+++ src/components/form/components/socialMedia/ui/SocialMedias.jsx
@@ -1,33 +1,12 @@
 import FormButton from "../../FormButton";
 import React, {useContext} from "react";
 import {ResumeContext} from "../../../../builder";
+import {handleSocialMedia} from "../units/handleSocialMedia";
+import {addSocialMedia} from "../units/addSocialMedia";
+import {removeSocialMedia} from "../units/removeSocialMedia";
 
 const SocialMedias = () => {
   const {resumeData, setResumeData} = useContext(ResumeContext);
-
-  // social media
-  const handleSocialMedia = (e, index) => {
-    const newSocialMedia = [...resumeData.socialMedia];
-    newSocialMedia[index][e.target.name] = e.target.value.replace(
-      "https://",
-      ""
-    );
-    setResumeData({...resumeData, socialMedia: newSocialMedia});
-  };
-
-  const addSocialMedia = () => {
-    setResumeData({
-      ...resumeData,
-      socialMedia: [...resumeData.socialMedia, {socialMedia: "", link: ""}],
-    });
-  };
-
-  const removeSocialMedia = (index) => {
-    const newSocialMedia = [...resumeData.socialMedia];
-    newSocialMedia[index] = newSocialMedia[newSocialMedia.length - 1];
-    newSocialMedia.pop();
-    setResumeData({...resumeData, socialMedia: newSocialMedia});
-  };
 
   return (
     <div className="flex-col-gap-2">
@@ -40,7 +19,7 @@ const SocialMedias = () => {
             name="socialMedia"
             className="other-input"
             value={socialMedia.socialMedia}
-            onChange={(e) => handleSocialMedia(e, index)}
+            onChange={(e) => handleSocialMedia(resumeData, setResumeData, e, index)}
           />
           <input
             type="text"
@@ -48,14 +27,18 @@ const SocialMedias = () => {
             name="link"
             className="other-input"
             value={socialMedia.link}
-            onChange={(e) => handleSocialMedia(e, index)}
+            onChange={(e) => handleSocialMedia(resumeData, setResumeData, e, index)}
           />
         </div>
       ))}
       <FormButton
         size={resumeData.socialMedia.length}
-        add={addSocialMedia}
-        remove={removeSocialMedia}
+        add={() => {
+          addSocialMedia(resumeData, setResumeData)
+        }}
+        remove={() => {
+          removeSocialMedia(resumeData, setResumeData)
+        }}
       />
     </div>
   );
