+++ src/components/form/components/socialMedia/units/addSocialMedia.js
@@ -1,0 +1,6 @@
+export const addSocialMedia = (resumeData, setResumeData) => {
+  setResumeData({
+    ...resumeData,
+    socialMedia: [...resumeData.socialMedia, {socialMedia: "", link: ""}],
+  });
+};
