+++ src/components/form/components/socialMedia/units/handleSocialMedia.js
@@ -1,0 +1,8 @@
+export const handleSocialMedia = (resumeData, setResumeData, e, index) => {
+  const newSocialMedia = [...resumeData.socialMedia];
+  newSocialMedia[index][e.target.name] = e.target.value.replace(
+    "https://",
+    ""
+  );
+  setResumeData({...resumeData, socialMedia: newSocialMedia});
+};
