+++ src/components/form/components/socialMedia/units/removeSocialMedia.js
@@ -1,0 +1,6 @@
+export const removeSocialMedia = (resumeData, setResumeData, index) => {
+  const newSocialMedia = [...resumeData.socialMedia];
+  newSocialMedia[index] = newSocialMedia[newSocialMedia.length - 1];
+  newSocialMedia.pop();
+  setResumeData({...resumeData, socialMedia: newSocialMedia});
+};
